requirements.txt
# ComfyUI启动器依赖包

# 核心依赖（Python标准库）
# tkinter - GUI框架（通常随Python安装）
# subprocess - 进程管理
# threading - 多线程支持
# pathlib - 路径操作
# json - JSON处理
# datetime - 日期时间处理
# hashlib - 哈希计算
# shutil - 文件操作

# 打包与运行依赖
pyinstaller>=5.0.0
PyQt5>=5.15.0
PyYAML>=6.0

# 可选依赖（用于扩展功能）
# GitPython==3.1.40  # Git操作（用于版本管理）
# psutil==5.9.6      # 系统监控（用于进程管理）
# requests==2.31.0   # HTTP请求（用于在线更新）

# 注意：
# 1. 启动器使用 PyQt5 作为唯一 GUI 框架
# 2. 构建可执行文件需要安装 pyinstaller
# 3. 若运行时出现导入错误，请检查 Python 与依赖安装是否完整
